--- packages/echogram/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@echo-webkom/echogram",
-  "version": "0.1.0",
-  "main": "./src/index.ts",
-  "scripts": {
-    "clean": "rm -rf .turbo node_modules",
-    "lint": "eslint src",
-    "lint:fix": "eslint lint --fix",
-    "typecheck": "tsc --noEmit"
-  },
-  "devDependencies": {
-    "eslint": "8.57.0",
-    "@types/node": "18.19.24",
-    "eslint-config-webkom": "workspace:*",
-    "typescript": "5.4.2"
-  },
-  "eslintConfig": {
-    "root": true,
-    "extends": [
-      "webkom/base"
-    ]
-  }
-}
